requirements_test.txt
pytest
pytest-asyncio
aiohttp
pytest-homeassistant-custom-component
ruff
